Removed no hawk's eye requirement on barrage
Logic. 60 potency loss if it happens to waste a proc of hawks.

In just bard buffs, getting barrage in the window is a 200  potency gain. Any drift allowed will eventually push it out of the window and cause it to miss out on the 200 potency, or even the full 1000 by skipping barrage all together.

Talking with others, mathed it, went with this.