Pitch perfect fix/ Sidewinder
Pitch perfect was requiring buffs to be running or on cd, was breaking with manual buff users who wouldn't buff during trash.

removed similar logic from Sidewinder as part of cleanup. Left it as part of the pooling option. Added its mention to pooling option text, it always did it, just adding the calrification.