Apex arrow small feature fix
Split the setting trying to replace burst shot and quick nock with apex arrow because it wasn't working for burst shot

Original setting became just aoe. Added new setting under parent combo of straight shot upgrade for st.

removed the code for turning it off with the turn off apex setting because that is for the one button mode not the small features.