Brd ogcd, more apex
Added heartbreakshot to ogcd feature for st so it works again

Moved the brd apex feature into the AOE combo feature just like with ST. It was redundant and doing it in both places.

removed unneeded bools in one button songs and buffs, using action ready

Cleared up conflicting combos throughout, moved not so simple modes to the top, we will be getting a real simple mode soon.